+++ app/src/main/java/com/dviltres/paymentapp/domain/repository/InstallmentRepository.kt
@@ -3,5 +3,5 @@ package com.dviltres.paymentapp.domain.repository
 import com.dviltres.paymentapp.domain.model.Installment
 
 interface InstallmentRepository {
-    suspend fun getInstallment(amount:Int, paymentMethodId:String, issuerId:String): Installment?
+    suspend fun getInstallment(amount:Double, paymentMethodId:String, issuerId:String): Installment?
 }
